+++ src/lib/components/Labels.ts
@@ -1,4 +1,9 @@
+import monoEmoij from './monoEmoji';
+
 export const CreateLabel = '+';
-export const DeleteLabel = '✖';
-export const ConfirmLabel = '✓';
-export const EditLabel = '✎';
+export const DeleteLabel = monoEmoij('✖');
+export const ConfirmLabel = monoEmoij('✓');
+export const EditLabel = monoEmoij('✎');
+export const PrivateLabel = '••••';
+export const EmptyLabel = '—';
+export const UnknownLabel = monoEmoij('🔒');